feat(wave3): IoT spatial reporting slice, hexon-path as-built reconcile, workspace clippy clean
- iot_spatial_reporting: plan.md authored; iot_reading table (anchor+petal
  scope, metric/value/units, RFC-3339 + epoch-ms + HLC timestamps, 3 indexes,
  no geometry by design) + all-or-nothing validated write handler with 7
  read-back tests; fe-query timeseries builders (latest_per_anchor,
  window_aggregate, anchors_within, readings_in_window; 8 rendered-SQL tests);
  authenticated batch ingestion POST /petals/:id/iot/readings (Editor+, scope
  + rate limits, 1000-row cap); iot_reading whitelisted in query_guard with
  scope-filter injection so /query + shared URLs serve IoT rows today. 24 tests.
- hexon_path_asset: spec reconciled to as-built PathAssetDescriptor; loader/
  P2P residual formally handed to hexon_p2p_bucket (cross-linked); FR-5/FR-6
  plan.md written for future pickup.
- code_review_20260430_clippy_quality (DONE): ~120 warnings + 5 errors fixed
  across the workspace (reviewed --fix + hand fixes for itemized issues #6,
  #7, #10, #14, #17; #11 measured + deferred); clippy --workspace
  --all-targets -D warnings green; cargo fmt green.

cargo test sweep deferred to session end per test policy.

+++ fe-webview/src/backends/tauri.rs
@@ -9,11 +9,11 @@ use wry::{PageLoadEvent, WebView, WebViewBuilder};
 
 use crate::backend::{BackendEvent, WebViewBackend, WindowGeometry};
 
-#[cfg(target_os = "windows")]
-use super::win32_popup::{win32, PopupHandle};
+use super::win32_popup::webview_fill_rect;
 #[cfg(not(target_os = "windows"))]
 use super::win32_popup::ParentHandle;
-use super::win32_popup::webview_fill_rect;
+#[cfg(target_os = "windows")]
+use super::win32_popup::{win32, PopupHandle};
 
 /// Tauri-powered webview backend; on Windows it lives in a borderless popup owned by the Bevy window.
 pub struct TauriBackend {
@@ -62,7 +62,10 @@ impl WebViewBackend for TauriBackend {
     ) -> anyhow::Result<Self> {
         tracing::info!(
             "TauriBackend::create — geometry: x={} y={} w={} h={}",
-            geometry.x, geometry.y, geometry.width, geometry.height
+            geometry.x,
+            geometry.y,
+            geometry.width,
+            geometry.height
         );
 
         let events: Rc<RefCell<Vec<BackendEvent>>> = Rc::new(RefCell::new(Vec::new()));
@@ -96,7 +99,7 @@ impl WebViewBackend for TauriBackend {
         #[cfg(target_os = "windows")]
         let (webview, popup_hwnd) = {
             let parent_hwnd = match parent_handle {
-                RawWindowHandle::Win32(h) => h.hwnd.get() as isize,
+                RawWindowHandle::Win32(h) => h.hwnd.get(),
                 _ => anyhow::bail!("TauriBackend on Windows requires a Win32 window handle"),
             };
 
@@ -107,13 +110,11 @@ impl WebViewBackend for TauriBackend {
 
             let popup = PopupHandle(popup_hwnd);
             tracing::debug!("TauriBackend: calling build_as_child...");
-            let webview = builder
-                .build_as_child(&popup)
-                .map_err(|e| {
-                    tracing::error!("TauriBackend: build_as_child FAILED: {e}");
-                    win32::destroy(popup_hwnd);
-                    anyhow::anyhow!("TauriBackend: build_as_child failed: {e}")
-                })?;
+            let webview = builder.build_as_child(&popup).map_err(|e| {
+                tracing::error!("TauriBackend: build_as_child FAILED: {e}");
+                win32::destroy(popup_hwnd);
+                anyhow::anyhow!("TauriBackend: build_as_child failed: {e}")
+            })?;
 
             tracing::info!("TauriBackend: webview built OK — hiding popup until navigate()");
             win32::hide(popup_hwnd);
